update organization for client side SOR logic
